requirements.txt
Flask==3.0.0
Flask-Login==0.6.3
Flask-WTF==1.2.1
Werkzeug==3.0.1
requests==2.31.0
PyJWT==2.12.1
gunicorn==21.2.0
Flask-Limiter==3.5.0
